Added the PDF Viewer Share Point sample
