+++ SharePoint/.npmignore
