+++ SharePoint/.vscode/settings.json
@@ -1,0 +1,1 @@
+{}
