+++ SharePoint/.yo-rc.json
