+++ SharePoint/config/deploy-azure-storage.json
