+++ SharePoint/config/write-manifests.json
