+++ SharePoint/src/webparts/pdfviewer/PdfviewerWebPart.manifest.json
