+++ SharePoint/src/webparts/pdfviewer/assets/welcome-dark.png
