+++ SharePoint/src/webparts/pdfviewer/loc/mystrings.d.ts
